+++ src/ast/irbuilderbpf.h
@@ -160,6 +160,7 @@ public:
                                const location &loc,
                                MDNode *metadata);
   CallInst *CreateGetNs(TimestampMode ts, const location &loc);
+  CallInst *CreateJiffies64(const location &loc);
   CallInst *CreateGetPidTgid(const location &loc);
   CallInst *CreateGetCurrentCgroupId(const location &loc);
   CallInst *CreateGetUidGid(const location &loc);