+++ tests/runtime/engine/runner.py
@@ -134,6 +134,7 @@ class Runner(object):
         bpffeature["skboutput"] = output.find("skboutput: yes") != -1
         bpffeature["get_tai_ns"] = output.find("get_ktime_ns: yes") != -1
         bpffeature["get_func_ip"] = output.find("get_func_ip: yes") != -1
+        bpffeature["jiffies64"] = output.find("jiffies64: yes") != -1
         return bpffeature
 
     @staticmethod